Delete Transpose
This seems to be working but need much more sophisticated work since it
gets really tricky with pointers and the incoming and outgoing Vertices
of each Vertex. I implemented it for Kosaraju’s Strongly Connected
Components but it turns out that I can just import the initial JSON
file in reverse way to construct a transpose graph. And this seems to
work fine for now, although Kosaraju algorithm implementation itself is
not stable for now.