Add error return value to GetNode() function
GetNode currently returns nil for a non-existent node, however other
functions (for example GetTargets) return an error for a non-existent
node. This patch adds an error return value to GetNode to make it
consistent with the rest of the package.

This patch is not a merge candidate because it includes import path
changes. These were necessary in order to run travis build but clearly
another solution is needed.